+++ doc/changelog.rst
@@ -4,7 +4,7 @@ Changelog
 Version 0.7.5
 -------------
 
-* Add ``Benchmark.add_metadata()`` method
+* Add ``Benchmark.update_metadata()`` method
 
 Version 0.7.4 (2018-08-18)
 --------------------------